Migrate Ajax search into Admin actions
